Configuration to support tls and generation of needed certs.
tls is enabled (not disabled) in vault.json with switch from
http to https in the Vault URL.

The gencerts script generates Vault CA and Server certs.  These
need to be placed in volumes/tls (ignored by git: must be created
locally -- but the init script sets up the directory).

README.md updated to reflect tls support and the coordination in
naming needed between Vault and Vault clients (esp. Concourse).